dev: Update CLI telemetry task definition
